% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_collaborations.R
\name{combine_collaborators}
\alias{combine_collaborators}
\title{Combine Collaborators}
\usage{
combine_collaborators(dataframe, id_col, collaborators)
}
\arguments{
\item{dataframe}{a dataframe object}

\item{id_col}{string. Name of column to be used as id.}

\item{collaborators}{vector of strings. the character vector with collaborators to be transformed into pairs.}
}
\value{
a dataframe of cleaned collaboration pairs with their id for later matching.
}
\description{
`combine_collaborators` takes an input dataframe with collaborations across columns,
subsets an id column and all specified collaboration columns, and returns a dataframe with three columns:
- id
- firm_1
- firm_2
}
\examples{
combine_collaborators(swinno_collaboration, "sinno_id", c("firm_1", "collaborator_1", ... "collaborator_n"))
}
