% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_usesectors.R
\name{filter_in_columns}
\alias{filter_in_columns}
\title{Filter Dataframe for Unique IDs Based on Filtered Values in Specific Columns}
\usage{
filter_in_columns(df, use_cols, id_col, filter_values)
}
\arguments{
\item{df}{The input dataframe to filter}

\item{use_cols}{A character vector indicating the range of columns to filter on}

\item{id_col}{A character string indicating the name of the ID column to filter on}

\item{filter_values}{A character vector containing the filter values to search for in the "use_sectors" column}
}
\value{
A vector of unique IDs that have values in the specified filter_values in the specified range of columns
}
\description{
This function takes in a dataframe, a range of columns to filter on, a specified ID column,
and a list of filter values. The function first converts the specified range of columns to character
if necessary, then melts the dataframe into a long format based on the range of columns, and
finally filters for unique IDs if the "use_sectors" column contains a value specified in the
filter_values list. This function requires the dplyr package to run.
}
\examples{
swinno
use_sectors_of_interest = c("123", "45")
filter_in_columns(swinno, use_cols=c("use_sectors", "use_sector_1", "use_sector_2"), "sinno_id", filter_values = use_sectors_of_interest)
}
