% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_usesectors.R
\name{melt_columns}
\alias{melt_columns}
\title{Melt specified columns of a dataframe into long format}
\usage{
melt_columns(df, use_cols, id_col, new_col_name)
}
\arguments{
\item{df}{A dataframe}

\item{use_cols}{A character vector specifying the range of columns to melt}

\item{id_col}{A character vector specifying the name of the column containing unique identifiers}
}
\value{
A dataframe with the specified columns melted into long format
}
\description{
This function takes a dataframe and a range of column names as inputs, and returns a new dataframe with the specified columns melted into long format.
}
\examples{
swinno
melt_columns(swinno, use_cols=c("use_sectors", "use_sector_1", "use_sector_2"), "sinno_id")
}
