% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_collaborations.R
\name{select_collaborators}
\alias{select_collaborators}
\title{Selects columns to include in collaboration pairs}
\usage{
select_collaborators(dataframe, id_col = "sinno_id", collaborators)
}
\arguments{
\item{dataframe}{the input dataframe}

\item{id_col}{string.The id column to be used, defaults to sinn_id}

\item{collaborators}{the character vector with collaborators to be transformed into pairs}
}
\value{
An object of the same type as the input. The output has the following properties:
* Columns are a subset of the input, but appear in the same order.
* Rows are not modified.
}
\description{
`select_collaborators` selects only the collaborators and an id column from swinno collaboration data.
}
\examples{
select_collaborators(swinno, "sinno_id", c("collaborator_1", "collaborator_2", "collaborator_3"))
}
