% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_swinno.R
\name{import_clean_swinno}
\alias{import_clean_swinno}
\title{Import SWINNO from Excel}
\usage{
import_clean_swinno(path)
}
\arguments{
\item{path}{the file path as a string.}
}
\value{
A dataframe with cleanly named SWINNO data.
}
\description{
`import_clean_swinno` imports SWINNO data from excel and cleans up the column
names to fit snakecase pattern. It returns a dataframe with names such as
sinno_id. The function takes additional arguments to specify different sheet
names than the default.
}
\examples{
import_clean_swinno('data/swinno.xslx')
}
